html/semantics/embedded-content-0/media-elements/currentSrc_property_exists.html

<!doctype html>
<html>
 <head>
   <title>{audio,video}.currentSrc - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script /resources/testharnessreport.js></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-currentsrc">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  "currentSrc" in document.getElementById("a"),
  "audioElement.currentSrc property exists");
}, "audio.currentSrc - existence test");

test(function() {
 assert_true(
  "currentSrc" in document.getElementById("v"),
  "audioElement.currentSrc property exists");
}, "video.currentSrc - existence test");
  </script>
 </body>
</html>
